updated image generator tool
