+++ app/services/schemas.py
@@ -163,34 +163,4 @@ class NoteGeneratorArgs(BaseModel):
     file_url: str
     lang: Optional[str] = "en"
 
-class OutlineGeneratorInput(BaseModel):
-    n_slides: int
-    topic: str
-    instructional_level: str
-    file_url: str
-    file_type: str
-    lang: Optional[str] = "en"
 
-class SlideGeneratorInput(BaseModel):
-    slides_titles: List[str]
-    instructional_level: str
-    topic: str
-    file_url: str
-    file_type: str
-    lang: Optional[str] = "en"
-
-# class ImageGeneratorInput(BaseModel):
-#     prompt: str
-#     grade_level: str
-#     subject: str
-#     presets: Optional[dict] = None
-#     lang: Optional[str] = "en"
-
-class ImageGeneratorArgs(BaseModel):
-    prompt: str
-    subject: Optional[str] = None
-    grade_level: Optional[str] = None
-    presets: Optional[dict] = None
-    size: Optional[str] = "1024x1024"
-    lang: Optional[str] = "en"   
-    style: Optional[str] = None
