+++ app/tools/image_generator/prompt/prompt-for-enhancement.txt
